requirements.txt
psycopg2-binary>=2.9,<3.0

